extracted generation steps
